+++ NEWS.md
@@ -12,6 +12,8 @@
 * Substantially faster performance by caching the granule cardinality graph 
   (used by `chronon_convert()`, `chronon_divmod()`, `chronon_common()` and
   `time_parts()`) and tzdb names (used in operations with timezones).
+* Reworked time vectors to be built with S7 rather than vctrs for improved
+  double dispatch methods for arithmetic and other operations (#63).
 
 # mixtime 0.2.0